Submission File Types, Category Management, Rubric Management, Minor additions of features

+++ app/client/PeerGrade/src/pages/classNav/Rubrics.jsx
@@ -6,9 +6,9 @@ import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@
 import { Badge } from "@/components/ui/badge";
 import { Card, CardHeader, CardTitle, CardDescription, CardContent, CardFooter } from "@/components/ui/card";
 import { useUser } from "@/contexts/contextHooks/useUser";
-import { Pencil, Trash2, Eye, FileText, Info, ChevronDown } from 'lucide-react';
+import { Pencil, Trash2, Eye, Info, ChevronDown } from 'lucide-react';
 import { deleteRubricsFromAssignment, getAllRubricsInClass, getRubricById } from '@/api/rubricApi';
-import { getAllAssignmentsByClassId } from '@/api/assignmentApi';
+import EditRubric from '@/components/rubrics/EditRubric';
 import {
   AlertDialog,
   AlertDialogTrigger,
@@ -27,50 +27,53 @@ const Rubrics = () => {
   const { classId } = useParams();
   const { user } = useUser();
   const [rubrics, setRubrics] = useState([]);
-  const [assignments, setAssignments] = useState([]);
   const [selectedRubric, setSelectedRubric] = useState(null);
   const [isDrawerOpen, setIsDrawerOpen] = useState(false);
-  const [isCreateDrawerOpen, setIsCreateDrawerOpen] = useState(false);
   const [isAlertDialogOpen, setIsAlertDialogOpen] = useState(false);
   const [rubricToDelete, setRubricToDelete] = useState(null);
+  const [isEditDrawerOpen, setIsEditDrawerOpen] = useState(false);
+  const [rubricToEdit, setRubricToEdit] = useState(null);
+  const [editDrawerKey, setEditDrawerKey] = useState(0); // Add this line
+  const [confirmDeleteRubric, setConfirmDeleteRubric] = useState(false);
 
   useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const [rubricsResponse, assignmentsResponse] = await Promise.all([
-          getAllRubricsInClass(classId),
-          getAllAssignmentsByClassId(classId)
-        ]);
-
-        if (rubricsResponse && rubricsResponse.status === 'Success' && Array.isArray(rubricsResponse.data)) {
-          setRubrics(rubricsResponse.data);
-        } else {
-          setRubrics([]);
-        }
-
-        if (assignmentsResponse && assignmentsResponse.data) {
-          setAssignments(assignmentsResponse.data);
-        } else {
-          setAssignments([]);
-        }
-      } catch (error) {
-        console.error('Error fetching data:', error);
-        setRubrics([]);
-        setAssignments([]);
-      }
-    };
-  
     fetchData();
   }, [classId]);
 
-  const handleDeleteRubric = async () => {
+  const fetchData = async () => {
     try {
-      await deleteRubricsFromAssignment(rubricToDelete.rubricId);
-      setIsDrawerOpen(false);
-      setIsAlertDialogOpen(false);
-      fetchRubrics();
+      const rubricsResponse = await getAllRubricsInClass(classId);
+      if (rubricsResponse && rubricsResponse.status === 'Success' && Array.isArray(rubricsResponse.data)) {
+        setRubrics(rubricsResponse.data);
+      } else {
+        setRubrics([]);
+      }
     } catch (error) {
-      console.error('Error deleting rubric:', error);
+      console.error('Error fetching rubrics:', error);
+      setRubrics([]);
+    }
+  };
+
+  const calculateTotalPoints = (rubric) => {
+    if (!rubric || !rubric.criteria) return 0;
+    return rubric.criteria.reduce((total, criterion) => 
+      total + Math.max(...criterion.criterionRatings.map(rating => parseFloat(rating.points) || 0)), 0
+    );
+  };
+
+  const handleDeleteRubric = async () => {
+    if (confirmDeleteRubric) {
+      setConfirmDeleteRubric(false);
+      try {
+        await deleteRubricsFromAssignment(rubricToDelete.rubricId);
+        await fetchData(); // Wait for the data to be fetched
+        setIsDrawerOpen(false);
+        setIsAlertDialogOpen(false);
+      } catch (error) {
+        console.error('Error deleting rubric:', error);
+      }
+    } else {
+      setConfirmDeleteRubric(true);
     }
   };
 
@@ -93,37 +96,50 @@ const Rubrics = () => {
     fetchRubrics();
   };
 
-  const fetchRubrics = async () => {
+  const handleEditRubricClick = async (rubricId) => {
     try {
-      const response = await getAllRubricsInClass(classId);
-      if (response && response.status === 'Success' && Array.isArray(response.data)) {
-        setRubrics(response.data);
-      } else {
-        setRubrics([]);
+      const response = await getRubricById(rubricId);
+      if (response.data && response.data) {
+        setRubricToEdit(response.data);
+        setIsEditDrawerOpen(true);
+        setEditDrawerKey(prevKey => prevKey + 1); // Add this line
       }
     } catch (error) {
-      console.error('Error fetching rubrics:', error);
-      setRubrics([]);
+      console.error('Error fetching rubric details:', error);
     }
+  };
+
+  const handleCloseEditDrawer = () => {
+    setIsEditDrawerOpen(false);
+    setRubricToEdit(null);
+  };
+
+  const handleRubricUpdated = (updatedRubric) => {
+    setRubrics(prevRubrics => prevRubrics.map(rubric => 
+      rubric.rubricId === updatedRubric.rubricId ? updatedRubric : rubric
+    ));
+    handleCloseEditDrawer(); // Close the drawer after updating
   };
 
   return (
     <div className="p-6 bg-muted rounded-lg">
       <h1 className="text-3xl font-bold mb-8 text-gray-800">Rubrics</h1>
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
+      <div className="flex flex-wrap gap-6">
         {rubrics.length === 0 ? (
           <div className="col-span-full text-center text-gray-500">No rubrics found</div>
         ) : (
           rubrics.map((rubric) => (
-            <Card key={rubric.rubricId} className="hover:shadow-md transition-shadow bg-white">
+            <Card key={rubric.rubricId} className="hover:shadow-md min-w-[300px] transition-shadow bg-white">
               <CardHeader className="pb-2 flex justify-between">
-                <div>
-                  <CardTitle className="text-xl font-semibold text-gray-800">{rubric.title}</CardTitle>
-                  <CardDescription className="text-sm text-gray-600 line-clamp-2">{rubric.description}</CardDescription>
-                </div>
-                <Badge variant="outline" className="text-xs w-1/2 flex justify-center items-center bg-success/30">
-                    {rubric.totalMarks} Points
-                </Badge>
+                 <div>
+                    <CardTitle className="text-xl font-semibold text-gray-800">{rubric.title}</CardTitle>
+                    <CardDescription className="text-sm text-gray-600 line-clamp-2">{rubric.description}</CardDescription>
+                  </div>
+                  <div>
+                    <Badge variant="outline" className="text-xs flex justify-center items-center bg-success/30">
+                      {rubric.totalMarks} Points
+                    </Badge>
+                  </div>
               </CardHeader>
               <CardContent className="pt-2">
               
@@ -131,10 +147,17 @@ const Rubrics = () => {
               <CardFooter>
                 <Button 
                   onClick={() => handleRubricClick(rubric.rubricId)} 
-                  className="w-full bg-gray-100 text-gray-800 hover:bg-gray-200"
+                  className="w-1/2 bg-gray-100 text-gray-800 hover:bg-gray-200 mr-2"
                   variant="ghost"
                 >
                   <Eye className="mr-2 h-4 w-4" /> View Details
+                </Button>
+                <Button 
+                  onClick={() => handleEditRubricClick(rubric.rubricId)} 
+                  className="w-1/2 bg-blue-100 text-blue-800 hover:bg-blue-200"
+                  variant="ghost"
+                >
+                  <Pencil className="mr-2 h-4 w-4" /> Edit Rubric
                 </Button>
               </CardFooter>
             </Card>
@@ -143,11 +166,41 @@ const Rubrics = () => {
       </div>
       <Drawer open={isDrawerOpen} onOpenChange={setIsDrawerOpen}>
         <DrawerContent>
-          <DrawerHeader>
-            <DrawerTitle className="text-2xl">{selectedRubric?.title}</DrawerTitle>
-            <DrawerDescription>{selectedRubric?.description}</DrawerDescription>
-            <div className="flex justify-end space-x-2 mt-4">
-              
+        <DrawerHeader>
+          <div className="flex justify-between items-start">
+            <div>
+              <DrawerTitle className="text-2xl">{selectedRubric?.title}</DrawerTitle>
+              <DrawerDescription>{selectedRubric?.description}</DrawerDescription>
+            </div>
+            <AlertDialog open={isAlertDialogOpen} onOpenChange={setIsAlertDialogOpen}>
+              <AlertDialogTrigger asChild>
+                <Button 
+                  onClick={() => {
+                    setRubricToDelete(selectedRubric);
+                    setConfirmDeleteRubric(false);
+                    setIsAlertDialogOpen(true);
+                  }} 
+                  variant="destructive"
+                  size="sm"
+                >
+                  <Trash2 className="h-4 w-4" /> <span className="ml-2">Delete Rubric</span>
+                </Button>
+              </AlertDialogTrigger>
+              <AlertDialogContent className={confirmDeleteRubric ? "text-white bg-red-500 border-red-800" : ""}>
+                <AlertDialogHeader>
+                  <AlertDialogTitle>{confirmDeleteRubric ? "Confirm " : ""}Delete Rubric</AlertDialogTitle>
+                  <AlertDialogDescription className={confirmDeleteRubric ? "text-white" : ""}>
+                    Are you {confirmDeleteRubric ? "really " : ""}sure you want to delete this rubric? This action cannot be undone.
+                  </AlertDialogDescription>
+                </AlertDialogHeader>
+                <AlertDialogFooter>
+                  <AlertDialogCancel asChild>
+                    <Button variant="outline" className={confirmDeleteRubric ? "text-black" : ""}>Cancel</Button>
+                  </AlertDialogCancel>
+                  <Button variant="destructive" onClick={handleDeleteRubric}>Delete</Button>
+                </AlertDialogFooter>
+              </AlertDialogContent>
+            </AlertDialog>
             </div>
           </DrawerHeader>
           <div className="p-4 bg-gray-100 rounded-md mb-4">
@@ -156,9 +209,9 @@ const Rubrics = () => {
               How to read this rubric:
             </h4>
             <ul className="text-sm space-y-1">
-              <li><Badge variant="outline" className="mr-2 bg-accent text-slate-900">Criteria</Badge> Specific aspects of the assignment being evaluated</li>
-              <li><Badge variant="outline" className="mr-2 bg-accent text-slate-900">Ratings</Badge> Descriptions of performance levels for each criterion</li>
-              <li><Badge variant="outline" className="mr-2 bg-accent text-slate-900">Points</Badge> Score associated with each rating level</li>
+              <li><Badge variant="outline" className="mr-2 bg-accent/70 text-slate-900">Criteria</Badge> Specific aspects of the assignment being evaluated</li>
+              <li><Badge variant="outline" className="mr-2 bg-accent/70 text-slate-900">Ratings</Badge> Descriptions of performance levels for each criterion</li>
+              <li><Badge variant="outline" className="mr-2 bg-accent/70 text-slate-900">Points</Badge> Score associated with each rating level</li>
             </ul>
           </div>
           <div className="px-4 pb-4 max-h-[60vh] overflow-y-auto">
@@ -186,50 +239,33 @@ const Rubrics = () => {
                     </TableCell>
                     <TableCell className="text-right">
                       <Badge variant="outline" className="bg-accent text-slate-900">
-                        {criterion.criterionRatings.reduce((sum, rating) => sum + (parseFloat(rating.points) || 0), 0)}
+                        {Math.max(...criterion.criterionRatings.map(rating => parseFloat(rating.points) || 0))}
                       </Badge>
                     </TableCell>
                   </TableRow>
                 ))}
-              </TableBody>
+            </TableBody>
             </Table>
-            <AlertDialog open={isAlertDialogOpen} onOpenChange={setIsAlertDialogOpen}>
-              <AlertDialogTrigger asChild>
-                <Button 
-                  onClick={() => {
-                    setRubricToDelete(selectedRubric);
-                    setIsAlertDialogOpen(true);
-                  }} 
-                  variant="destructive"
-                >
-                  <Trash2 className="mr-2 h-4 w-4" /> Delete Rubric
-                </Button>
-              </AlertDialogTrigger>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>Confirm Delete</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    Are you sure you want to delete this rubric? This action cannot be undone.
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <AlertDialogCancel asChild>
-                    <Button variant="outline">Cancel</Button>
-                  </AlertDialogCancel>
-                  <AlertDialogAction asChild>
-                    <Button variant="destructive" onClick={handleDeleteRubric}>Delete</Button>
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
+            
           </div>
-          <DrawerFooter>
-            <p className="text-sm text-gray-500">Total Points: {selectedRubric?.totalMarks}</p>
+         <DrawerFooter>
+            <div className="w-full flex justify-end">
+              <p className="text-sm text-gray-500">
+                Total Points: {calculateTotalPoints(selectedRubric)}
+              </p>
+            </div>
             <Button variant="ghost" onClick={() => setIsDrawerOpen(false)}><ChevronDown className='w-6 h-6'/></Button>
           </DrawerFooter>
         </DrawerContent>
       </Drawer>
 
+      <EditRubric 
+        key={editDrawerKey} // Add this line
+        isOpen={isEditDrawerOpen}
+        onClose={handleCloseEditDrawer}
+        rubricData={rubricToEdit}
+        onRubricUpdated={handleRubricUpdated}
+      />
     
     </div>
   );
